@@ -4,52 +4,46 @@
   "dependencies": {
     "@hotosm/iso-countries-languages": "^1.1.2",
     "@karaokemugen/i18n-iso-languages": "^3.1.1",
-    "@sentry/integrations": "^6.17.6",
-    "@sentry/react": "^6.17.6",
-    "antd": "4.18.6",
-    "dompurify": "^2.3.5",
-    "i18next": "^21.6.11",
-    "lodash.debounce": "^4.0.8",
-    "lodash.difference": "^4.5.0",
-    "lodash.sample": "^4.2.1",
-    "lodash.shuffle": "^4.2.0",
+    "@sentry/integrations": "^6.18.1",
+    "@sentry/react": "^6.18.1",
+    "antd": "4.18.9",
+    "dompurify": "^2.3.6",
+    "i18next": "^21.6.13",
+    "lodash": "^4.17.21",
     "minireset.css": "^0.0.7",
     "path-browserify": "^1.0.1",
-    "pretty-bytes": "^5.6.0",
+    "pretty-bytes": "^6.0.0",
     "react": "17.0.2",
     "react-aria-menubutton": "^7.0.3",
     "react-beautiful-dnd": "^13.1.0",
     "react-dom": "17.0.2",
-    "react-i18next": "^11.15.4",
+    "react-i18next": "^11.15.5",
     "react-image-crop": "^9.1.1",
-    "react-router-dom": "6.2.1",
+    "react-router-dom": "6.2.2",
     "react-text-transition": "^1.3.0",
-    "react-toastify": "^8.1.1",
-    "react-virtuoso": "^2.6.0",
+    "react-toastify": "^8.2.0",
+    "react-virtuoso": "^2.7.2",
     "resize-observer-polyfill": "^1.5.1",
-    "sass": "^1.49.7",
+    "sass": "^1.49.9",
     "slugify": "^1.6.5",
     "socket.io-client": "4.4.1",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
     "@fortawesome/fontawesome-free": "^6.0.0",
-    "@types/lodash.debounce": "^4.0.6",
-    "@types/lodash.difference": "^4.5.6",
-    "@types/lodash.sample": "^4.2.6",
-    "@types/lodash.shuffle": "^4.2.6",
+    "@types/lodash": "^4.14.179",
     "@types/react": "17.0.39",
     "@types/react-aria-menubutton": "^6.2.9",
     "@types/react-beautiful-dnd": "^13.1.2",
-    "@types/react-dom": "17.0.11",
-    "eslint-config-prettier": "^8.3.0",
+    "@types/react-dom": "17.0.13",
+    "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-security": "^1.4.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
     "eslint-plugin-you-dont-need-lodash-underscore": "^6.12.0",
     "i18next-scanner": "^3.1.0",
     "react-scripts": "5.0.0",
     "source-map-explorer": "^2.5.2",
-    "typescript": "4.5.5"
+    "typescript": "4.6.2"
   },
   "scripts": {
     "analyze": "source-map-explorer 'build/static/js/*.js'",
