@@ -193,24 +176,25 @@
     "@typescript-eslint/eslint-plugin": "5",
     "@typescript-eslint/parser": "5",
     "chai": "^4.3.6",
-    "electron": "^17.0.0",
-    "eslint": "8",
+    "electron": "^17.1.0",
+    "esbuild": "^0.14.23",
+    "eslint": "^8.10.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^16.1.0",
-    "eslint-config-prettier": "^8.3.0",
+    "eslint-config-prettier": "^8.4.0",
     "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-security": "^1.4.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
     "husky": "^7.0.4",
     "i18next-scanner": "^3.1.0",
-    "lint-staged": "^12.3.3",
-    "mocha": "^9.2.0",
-    "postgrator-cli": "^4.0.0",
+    "lint-staged": "^12.3.4",
+    "mocha": "^9.2.1",
+    "postgrator-cli": "^5.1.0",
     "prettier": "2.5.1",
     "rimraf": "^3.0.2",
     "ts-node": "^10.5.0",
-    "typescript": "^4.5.5"
+    "typescript": "^4.6.2"
   },
   "build": {
     "appId": "KaraokeMugen",
