@@ -1,17 +1,17 @@
 const kpath = 'app/karaokebase/karaokes';
 const parentsFile = 'parents.yaml';
 
-const yaml = require('js-yaml');
-const fs = require('fs');
-const path = require('path');
+import yaml from 'js-yaml';
+import { readFileSync, writeFileSync } from 'fs';
+import * as path from 'path';
 
-const parentsData = fs.readFileSync(parentsFile, 'utf-8');
+const parentsData = readFileSync(parentsFile, 'utf-8');
 const p = yaml.load(parentsData);
 
 for (const kara of p.parents) {
 	let karaParentData = '';
 	try {
-		karaParentData = fs.readFileSync(path.resolve(kpath, kara.karaParent));
+		karaParentData = readFileSync(path.resolve(kpath, kara.karaParent));
 	} catch (err) {
 		console.log(kara.karaParent, 'not found');
 		continue;
@@ -21,7 +21,7 @@ for (const kara of p.parents) {
 	for (const child of kara.sons) {
 		let karaChildData = '';
 		try {
-			karaChildData = fs.readFileSync(path.resolve(kpath, child));
+			karaChildData = readFileSync(path.resolve(kpath, child));
 		} catch (err) {
 			console.log(child, 'not found');
 			continue;
@@ -39,6 +39,6 @@ for (const kara of p.parents) {
 				}, {});
 			karaChild.data = { ...ordered };
 		}
-		fs.writeFileSync(path.resolve(kpath, child), JSON.stringify(karaChild, null, 2));
+		writeFileSync(path.resolve(kpath, child), JSON.stringify(karaChild, null, 2));
 	}
 }
