@@ -1,17 +1,17 @@
 // This script changes a user's password
 
-const { Pool } = require('pg');
-const { hash, genSalt } = require('bcrypt');
-const { load } = require('js-yaml');
-const { readFileSync } = require('fs');
+import pg from 'pg';
+import bcrypt from 'bcryptjs';
+import { load } from 'js-yaml';
+import { readFileSync } from 'fs';
 
-async function hashPassword(password: string): Promise<string> {
-	return hash(password, await genSalt(10));
+async function hashPassword(password) {
+	return bcrypt.hash(password, 10);
 }
 
 async function main() {
 	const configFile = readFileSync('app/config.yml', 'utf-8');
-	const config: any = load(configFile);
+	const config = load(configFile);
 	const dbConfig = {
 		host: config.System.Database.host,
 		user: config.System.Database.username,
@@ -19,7 +19,7 @@ async function main() {
 		password: config.System.Database.password,
 		database: config.System.Database.database,
 	};
-	const client = new Pool(dbConfig);
+	const client = new pg.Pool(dbConfig);
 	const user = process.argv[2];
 	const password = await hashPassword(process.argv[3]);
 	try {
