@@ -16,10 +16,11 @@ export function initAutoUpdate() {
 	autoUpdater.autoDownload = false;
 	autoUpdater.on('error', error => {
 		logger.error('', { service: 'AppUpdate', obj: error });
+		const errMsg = error === null ? 'unknown' : (error.stack || error).toString();
 		dialog.showMessageBox({
 			type: 'none',
 			title: i18next.t('ERROR'),
-			message: error === null ? 'unknown' : (error.stack || error).toString(),
+			message: `${i18next.t('UPDATE_CHECK_ERROR')} : ${errMsg}`,
 		});
 	});
 	autoUpdater.on('update-available', async () => {
