@@ -1,5 +1,5 @@
 import i18n from 'i18next';
-import shuffle from 'lodash.shuffle';
+import { shuffle } from 'lodash';
 
 import { TipsAndTricks, TipType } from '../types/tips';
 
