@@ -1,7 +1,7 @@
 /**
  * How to use
  *
- * ts-node socketClient.ts
+ * ts-node socketClient.js
  * in a TTY-compatible terminal (cmd on Windows, not git bash)
  * use sendCommand('login', {body: {...}})  with the right params
  *
@@ -12,7 +12,7 @@ import { io } from 'socket.io-client';
 
 const socket = io(process.argv[2] || 'http://localhost:1337');
 
-function sendCommand(name: string, data: any) {
+function sendCommand(name, data) {
 	socket.emit(name, data, ack => {
 		console.log(JSON.stringify(ack, null, 2));
 	});
